refactor(usage-limit): move period helpers out of usage module
UsagePeriodService took a usage-limit type while usage-limit already imports UsageModule. The service, the UsagePeriod type and the calendar period utils are only consumed by usage-limit, so they live in usage-limit now.
