+++ packages/twenty-server/src/engine/core-modules/usage-limit/interfaces/credit-allowance-provider.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 import { type CreditAllowance } from 'src/engine/core-modules/usage-limit/types/credit-allowance.type';
-import { type UsagePeriod } from 'src/engine/core-modules/usage/types/usage-period.type';
+import { type UsagePeriod } from 'src/engine/core-modules/usage-limit/types/usage-period.type';
 
 @Injectable()
 export abstract class CreditAllowanceProvider {
