+++ packages/twenty-server/src/engine/core-modules/usage-limit/types/credit-allowance.type.ts
@@ -1,4 +1,4 @@
-import { type UsagePeriod } from 'src/engine/core-modules/usage/types/usage-period.type';
+import { type UsagePeriod } from 'src/engine/core-modules/usage-limit/types/usage-period.type';
 
 export type CreditAllowance = UsagePeriod & {
   allowanceMicro: number;
