+++ packages/twenty-server/src/engine/core-modules/usage-limit/utils/get-calendar-month-period.util.ts
@@ -1,6 +1,6 @@
 /* @license Enterprise */
 
-import { type UsagePeriod } from 'src/engine/core-modules/usage/types/usage-period.type';
+import { type UsagePeriod } from 'src/engine/core-modules/usage-limit/types/usage-period.type';
 
 export const getCalendarMonthPeriod = (now: Date): UsagePeriod => ({
   periodStart: new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), 1)),
